+++ wemake_python_styleguide/logics/variables.py
@@ -2,6 +2,7 @@
 
 from typing import Iterable, Optional
 
+from wemake_python_styleguide import constants
 from wemake_python_styleguide.options.defaults import MIN_VARIABLE_LENGTH
 
 
@@ -11,14 +12,19 @@ def is_wrong_variable_name(name: str, to_check: Iterable[str]) -> bool:
 
     >>> is_wrong_variable_name('wrong', ['wrong'])
     True
+
     >>> is_wrong_variable_name('correct', ['wrong'])
     False
+
     >>> is_wrong_variable_name('_wrong', ['wrong'])
     True
+
     >>> is_wrong_variable_name('wrong_', ['wrong'])
     True
+
     >>> is_wrong_variable_name('wrong__', ['wrong'])
     False
+
     >>> is_wrong_variable_name('__wrong', ['wrong'])
     False
 
@@ -43,14 +49,19 @@ def is_too_short_variable_name(
 
     >>> is_too_short_variable_name('test')
     False
+
     >>> is_too_short_variable_name(None)
     False
+
     >>> is_too_short_variable_name('o')
     True
+
     >>> is_too_short_variable_name('_')
     False
+
     >>> is_too_short_variable_name('z1')
     False
+
     >>> is_too_short_variable_name('z', min_length=1)
     False
 
@@ -64,12 +75,16 @@ def is_private_variable(name: Optional[str]) -> bool:
 
     >>> is_private_variable(None)
     False
+
     >>> is_private_variable('regular')
     False
+
     >>> is_private_variable('__private')
     True
+
     >>> is_private_variable('_protected')
     False
+
     >>> is_private_variable('__magic__')
     False
 
@@ -77,3 +92,36 @@ def is_private_variable(name: Optional[str]) -> bool:
     return (
         name is not None and name.startswith('__') and not name.endswith('__')
     )
+
+
+def is_variable_name_with_underscored_number(name: str) -> bool:
+    """
+    Checks for variable names with underscored number.
+
+    >>> is_variable_name_with_underscored_number('star_wars_episode2')
+    False
+
+    >>> is_variable_name_with_underscored_number(None)
+    False
+
+    >>> is_variable_name_with_underscored_number('come2_me')
+    False
+
+    >>> is_variable_name_with_underscored_number('_')
+    False
+
+    >>> is_variable_name_with_underscored_number('z1')
+    False
+
+    >>> is_variable_name_with_underscored_number('star_wars_episode_2')
+    True
+
+    >>> is_variable_name_with_underscored_number('come_2_me')
+    True
+
+    >>> is_variable_name_with_underscored_number('iso_123_456')
+    False
+
+    """
+    pattern = constants.UNDERSCORED_NUMBER_PATTERN
+    return name is not None and pattern.match(name) is not None