+++ wemake_python_styleguide/violations/naming.py
@@ -124,6 +124,7 @@ Summary
    TooShortVariableNameViolation
    PrivateNameViolation
    SameAliasImportViolation
+   UnderScoredNumberNameViolation
 
 Module names
 ------------
@@ -141,6 +142,7 @@ Variable names
 .. autoclass:: TooShortVariableNameViolation
 .. autoclass:: PrivateNameViolation
 .. autoclass:: SameAliasImportViolation
+.. autoclass:: UnderScoredNumberNameViolation
 
 """
 
@@ -439,3 +441,38 @@ class SameAliasImportViolation(ASTViolation):
     #: Error message shown to the user.
     error_template = 'Found same alias import "{0}"'
     code = 113
+
+
+class UnderScoredNumberNameViolation(SimpleViolation):
+    """
+    Forbids to have names with underscored numbers pattern.
+
+    Reasoning:
+        This is done for consistency in naming.
+
+    Solution:
+        Do not put an underscore between text and numbers, that confusing.
+        Rename your variable or modules to not include underscored numbers.
+
+    This rule checks: variables, and modules.
+    Please, note that putting an underscore that replaces ``-`` in some
+    names between numbers is fine, example: ``ISO-123-456`` would became
+    ``iso_123_456``.
+
+    Example::
+
+        # Correct:
+        star_wars_episode2 = 'awesome!'
+        iso_123_456 = 'some data'
+
+        # Wrong:
+        star_wars_episode_2 = 'not so awesome'
+
+    Note:
+        Returns Z114 as error code
+
+    """
+
+    #: Error message shown to the user.
+    error_template = 'Found underscored name pattern "{0}"'
+    code = 114