+++ wemake_python_styleguide/visitors/presets/general.py
@@ -1,15 +1,19 @@
 # -*- coding: utf-8 -*-
 
-from wemake_python_styleguide.visitors.ast import keywords, naming, numbers
-from wemake_python_styleguide.visitors.ast.classes import WrongClassVisitor
-from wemake_python_styleguide.visitors.ast.comparisons import (
-    ConstantComparisonVisitor,
+from wemake_python_styleguide.visitors.ast import (
+    comparisons,
+    keywords,
+    naming,
+    numbers,
 )
+from wemake_python_styleguide.visitors.ast.classes import WrongClassVisitor
 from wemake_python_styleguide.visitors.ast.functions import (
     WrongFunctionCallVisitor,
 )
 from wemake_python_styleguide.visitors.ast.imports import WrongImportVisitor
-from wemake_python_styleguide.visitors.ast.modules import WrongContentsVisitor
+from wemake_python_styleguide.visitors.ast.modules import (
+    EmptyModuleContentsVisitor,
+)
 from wemake_python_styleguide.visitors.ast.strings import WrongStringVisitor
 from wemake_python_styleguide.visitors.filenames.wrong_module_name import (
     WrongModuleNameVisitor,
@@ -20,18 +24,26 @@ GENERAL_PRESET = (
     # General:
     keywords.WrongRaiseVisitor,
     keywords.WrongKeywordVisitor,
+    keywords.WrongListComprehensionVisitor,
+    keywords.WrongForElseVisitor,
+
     WrongFunctionCallVisitor,
     WrongImportVisitor,
+
     naming.WrongNameVisitor,
     naming.WrongModuleMetadataVisitor,
+
     numbers.MagicNumberVisitor,
     WrongStringVisitor,
-    WrongContentsVisitor,
-    ConstantComparisonVisitor,
+
+    comparisons.ConstantComparisonVisitor,
+    comparisons.WrongOrderVisitor,
+    comparisons.MultipleInVisitor,
 
     # Classes:
     WrongClassVisitor,
 
     # Modules:
     WrongModuleNameVisitor,
+    EmptyModuleContentsVisitor,
 )